phasing out single JobStatus enum for separate BatchJobStatus and StreamingJobStatus enums
